Move ConnectionState to its own file to fix bundling issues (#10477)
This PR moves the `ConnectionState` enum to its own file within the container-loader package. This is done to prevent unexpectedly large bundle sizes when only importing `ConnectionState` from container-loader.
