+++ examples/utils/bundle-size-tests/webpack.config.js
@@ -12,6 +12,7 @@ const { BannedModulesPlugin } = require('@fluidframework/bundle-size-tools')
 module.exports = {
   entry: {
     'aqueduct': './src/aqueduct',
+    'connectionState': './src/connectionState',
     'containerRuntime': './src/containerRuntime',
     'loader': './src/loader',
     'map': './src/map',
